backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
django-redis==5.4.0
django-cors-headers==4.3.1
locust==2.17.0
